PyStationB/libraries/GlobalPenalisation/requirements.txt
azureml-core
botorch==0.5.0
gpy==1.10
gpytorch==1.5.0
jax==0.2.13
jaxlib==0.1.67
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
pyyaml==5.4.1
scipy==1.6.3
seaborn==0.11.1
setuptools=52.0.0
six==1.16.0
torch==1.8.1
tqdm==4.60.0
wheel=0.36.2
git+git://github.com/BrunoKM/emukit@gradients-for-covariance-gpy#egg=emukit
  